requirements.txt
whisperx @ git+https://github.com/m-bain/whisperx.git@befe2b242eb59dcd7a8a122d127614d5c63d36e9
python-dotenv==1.0.0
yt-dlp==2023.10.13
boto3==1.28.78
torchvision==0.16.1
